only member of a challenge can delete MR

+++ app/controllers/membership_readings_controller.rb
@@ -35,22 +35,27 @@ class MembershipReadingsController < ApplicationController
 
   def destroy
     @membership = membership_reading.membership
+    @challenge = @membership.challenge
+    if @challenge.has_member?(current_user)
 
-    membership_reading.destroy  # this needs to only destroy membershipreadings the user owns!! write test todo
+      membership_reading.destroy
 
-    current_user.delay.update_stats  #needs to be backgrounded!
-    @membership.delay.update_stats  #needs to be backgrounded!
-    @membership.group.delay.update_stats if @membership.group  #needs to be backgrounded!
+      current_user.delay.update_stats  #needs to be backgrounded!
+      @membership.delay.update_stats  #needs to be backgrounded!
+      @membership.group.delay.update_stats if @membership.group  #needs to be backgrounded!
 
-    respond_to do |format|
-      format.html { 
-        # go back to referer unless alternate location passed in
-        redirect = params[:location] || request.referer
-        # might be an anchor tag
-        redirect += params[:anchor] if params[:anchor]
-        redirect_to redirect
-      }
-      format.json { head :no_content }
+      respond_to do |format|
+        format.html { 
+          # go back to referer unless alternate location passed in
+          redirect = params[:location] || request.referer
+          # might be an anchor tag
+          redirect += params[:anchor] if params[:anchor]
+          redirect_to redirect
+        }
+        format.json { head :no_content }
+      end
+    else
+      raise "Not allowed"
     end
   end
